add total float example
