+++ src/float.rs
@@ -32,6 +32,60 @@
     include_str!("../examples/float_maps.rs"),
     "\n````"
 )]
+//!
+//! ## Total Float Example
+//!
+//! Unlike `Finite`, `Total` uses `total_cmp` order, so it can also represent
+//! NaN, the infinities, and -0.0 as distinct, orderable values. This example
+//! builds a category map over all of `f64` from nothing but
+//! `TotalF64::MIN`/`MAX`, the two infinities, and `-0.0`: start with NaN
+//! everywhere, carve out the finite range between the infinities as Normal,
+//! then punch in the infinities and -0.0 as exceptions (overlapping ranges
+//! have right-to-left precedence, so later entries win).
+//!
+//!```
+//! use range_set_blaze::{RangeMapBlaze, TotalF64, total::tf64};
+//!
+//! #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+//! enum Category {
+//!     NaN,
+//!     NegInfinity,
+//!     PosInfinity,
+//!     MinusZero,
+//!     Normal,
+//! }
+//!
+//! let category_map = RangeMapBlaze::from_iter([
+//!     (TotalF64::MIN..=TotalF64::MAX, Category::NaN),
+//!     (tf64(f64::NEG_INFINITY)..=tf64(f64::INFINITY), Category::Normal),
+//!     (tf64(f64::NEG_INFINITY)..=tf64(f64::NEG_INFINITY), Category::NegInfinity),
+//!     (tf64(f64::INFINITY)..=tf64(f64::INFINITY), Category::PosInfinity),
+//!     (tf64(-0.0)..=tf64(-0.0), Category::MinusZero),
+//! ]);
+//!
+//! # assert_eq!(category_map.get(tf64(f64::NAN)), Some(&Category::NaN));
+//! # assert_eq!(category_map.get(tf64(f64::NEG_INFINITY)), Some(&Category::NegInfinity));
+//! # assert_eq!(category_map.get(tf64(f64::INFINITY)), Some(&Category::PosInfinity));
+//! # assert_eq!(category_map.get(tf64(-0.0)), Some(&Category::MinusZero));
+//! # assert_eq!(category_map.get(tf64(0.0)), Some(&Category::Normal));
+//!
+//! for (range, category) in category_map.range_values() {
+//!     let (start, end) = (range.start().into_inner(), range.end().into_inner());
+//!     println!(
+//!         "{start:e} (0x{:016x}) ..= {end:e} (0x{:016x}) -> {category:?}",
+//!         start.to_bits(),
+//!         end.to_bits(),
+//!     );
+//! }
+//! // Output:
+//! // NaN (0xffffffffffffffff) ..= NaN (0xfff0000000000001) -> NaN
+//! // -inf (0xfff0000000000000) ..= -inf (0xfff0000000000000) -> NegInfinity
+//! // -1.7976931348623157e308 (0xffefffffffffffff) ..= -5e-324 (0x8000000000000001) -> Normal
+//! // -0e0 (0x8000000000000000) ..= -0e0 (0x8000000000000000) -> MinusZero
+//! // 0e0 (0x0000000000000000) ..= 1.7976931348623157e308 (0x7fefffffffffffff) -> Normal
+//! // inf (0x7ff0000000000000) ..= inf (0x7ff0000000000000) -> PosInfinity
+//! // NaN (0x7ff0000000000001) ..= NaN (0x7fffffffffffffff) -> NaN
+//!```
 
 #[doc(hidden)]
 pub mod finite_float;
